Fix Kodowania
NET wykrywał kodowanie UTF-8 w skryptach, gdzie własciwie kodowanie to
ANSI, przez co parser pluł błędami przy kompilacji, jeśli ktoś już
zmienił kodowanie pewnych plików może je przywrócić poprzez notepad++
(Format->Koduj w ANSII) lub poprzez program ponownie zamianiając
instance itema na taką sama instancje jaka ma np.
(ITMINUGGET=>ITMINUGGET)

+++ GothicItemsLookup/Results/resultsManager.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using LogSys;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -15,10 +16,12 @@ namespace GothicItemsLookup.Results
     {
         static public void deleteResultsFiles(ListBox.SelectedObjectCollection inResults)
         {
+            if (inResults == null) return;
             foreach (searchResult res in inResults)
             {
                 try
                 {
+                    new LogMsg("Usuwanie pliku: " + res.src.file, eDebugMsgLvl.INFO);
                     System.IO.File.Delete(res.src.file);
                     inResults.Remove(res);
                 }
